refactor: extract version package and display version in UI header
- Created new pkg/version package with GetVersion() function and version command
- Moved version.go from myhome/ to pkg/version/ and changed package name
- Exported GetVersion() function (previously getVersion()) for external use
- Updated CommandLineContext() to call version.GetVersion() instead of accepting version parameter
- Removed Version parameter from CommandLineContext() calls in main.go, ctl.go, and datacollector
- Added
