+++ go.work
@@ -1,4 +1,4 @@
-go 1.24.2
+go 1.25.3
 
 use (
 	.
@@ -52,4 +52,5 @@ use (
 	./pkg/shelly/types
 	./pkg/shelly/wifi
 	./pkg/tapo
+	./pkg/version
 )
