+++ internal/myhome/ui/static/index.html
@@ -26,7 +26,7 @@
         <div class="level">
           <div class="level-left">
             <h1 class="title is-3">MyHome</h1>
-            <span class="subtitle is-6 ml-3">Known devices ({{len .Devices}})</span>
+            <span class="subtitle is-6 ml-3">{{len .Devices}} devices / {{.Version}}</span>
           </div>
           <div class="level-right">
             <a class="button is-info" href="https://control.shelly.cloud" target="_blank" rel="noopener noreferrer">Shelly Control</a>
