+++ myhome/daemon/daemon.go
@@ -8,11 +8,11 @@ import (
 	"myhome/devices/impl"
 	mqttclient "myhome/mqtt"
 	mqttserver "myhome/mqtt"
+	mynet "myhome/net"
 	"myhome/occupancy"
 	"myhome/storage"
 	"myhome/temperature"
 	"myhome/ui"
-	"myhome/net"
 	"net/http"
 	_ "net/http/pprof"
 	"pkg/shelly"
