+++ pkg/version/go.mod
@@ -1,0 +1,10 @@
+module version
+
+go 1.25.3
+
+require github.com/spf13/cobra v1.10.1
+
+require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/pflag v1.0.10 // indirect
+)
